FIRM28: Refactor combined GPIO library into separate files
